Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Could you update with other committer's name as per the project proposal - https://wiki.opendaylight.org/display/ODL/ODL+SAF+Project
> Thanks,
> Prem
> 
> Following people needs to be added to committer's list.
> 
> Jeff Hartley - jhartley@luminanetworks.com
> Mahesh L - mahesh.lokhande@luminanetworks.com
> Sowmiya P - spichandi@luminanetworks.com
> Pravin - pdamodaran@luminanetworks.com
> Kangasundaram - kks@luminanetworks.com

Hi Prem: This INFO yaml is a autogenerated file read from ldap. It looks like these emails are not associated with the LFID's. While creating the project I have sent them invitations to the emails lists on the project wiki. Pls provide the correct LFID/emails to be updated here.

Patch-set: 1